add scroll state hook
