+++ src/nft/hooks/index.ts
@@ -11,6 +11,7 @@ export * from './useProfilePageState'
 export * from './useSelectAsset'
 export * from './useSellAsset'
 export * from './useSendTransaction'
+export * from './useSubscribeScrollState'
 export * from './useSweep'
 export * from './useTransactionResponse'
 export * from './useWalletBalance'
